Hw11 ready

+++ src/main/java/hw/homework11/Animal.java
@@ -2,11 +2,11 @@ package hw.homework11;
 
 class Animal {
     protected static String name;
-    protected static double runningSpeed;
-    protected static double swimmingSpeed;
+    protected static int runningSpeed;
+    protected static int swimmingSpeed;
     protected static int endurance;
 
-    public Animal(String name, double runningSpeed, double swimmingSpeed, int endurance) {
+    public Animal(String name, int runningSpeed, int swimmingSpeed, int endurance) {
         this.name = name;
         this.runningSpeed = runningSpeed;
         this.swimmingSpeed = swimmingSpeed;
