+++ src/main/java/hw/homework11/Dog.java
@@ -1,11 +1,11 @@
 package hw.homework11;
 
 class Dog extends Animal {
-    public Dog(String name, double runningSpeed, double swimmingSpeed, int endurance) {
+    public Dog(String name, int runningSpeed, int swimmingSpeed, int endurance) {
         super(name, runningSpeed, swimmingSpeed, endurance);
     }
 
-    public static double run(int distance) {
+    public static int run(int distance) {
         if (distance <= 0) {
             System.out.println("Недопустимое расстояние. Расстояние должно быть больше 0.");
             return -1;
@@ -16,13 +16,13 @@ class Dog extends Animal {
             return -1;
         }
 
-        double time = distance / runningSpeed;
+        int time = distance / runningSpeed;
         endurance -= distance;
         System.out.println(name + " бежал " + distance + " метров " + time + " секунд.");
         return time;
     }
 
-    public static double swim(int distance) {
+    public static int swim(int distance) {
         if (distance <= 0) {
             System.out.println("Недопустимое расстояние. Расстояние должно быть больше 0.");
             return -1;
@@ -33,7 +33,7 @@ class Dog extends Animal {
             return -1;
         }
 
-        double time = distance / swimmingSpeed;
+        int time = distance / swimmingSpeed;
         endurance -= distance * 2;
         System.out.println(name + " плавал " + distance + " метров " + time + " секунд.");
         return time;
